Add files via upload
The dashboard shows that while there's a good number of customers, the rate at which they leave (churn) is higher than the target. The analysis suggests that customers in the 31-50 age range, those with credit scores between 601-700, and those with account balances between 10,000 and 100,000 are likelier to leave. Delhi has the most customers, and there's a slight skew towards more female customers. A large portion of customers have credit cards, but a concerning number are currently inactive.

This information could be used to understand which customer segments are most at risk of leaving and to develop strategies to reduce churn in those specific groups.